🐛 Fix interactions with the remaining objects
Shelf, Tree and Boulder are on the 8th layer (Walls) and because of that, the player can't "see" them because the objects are blocking themselves. This commit changes the objects' layer just for a moment when it's checked if the player can see them.
LockdownWall has a toggleable sprite hitbox, which can be just left turned on.